Add source code string for emojis
Right now we only had the unicode character in the JSON file but depending on your editor, it's sometimes hard to read the source code. This commit keeps the current emoji in an `emojiChar` property (it helps to find an emoji in the file sometimes!) and adds the "source code" version in the `emoji` property.