Support emojis from Unicode 10.0
Adding support for all the emojis added in Unicode 10.0. See https://emojipedia.org/unicode-10.0/ for the list.
I also added a yet-not-supported `supports_gender` property in case we want to use it later.

I also had to modify the `cricket` alias to remove it from the `cricket bat and ball` emoji to the `cricket` (bug) emoji. This is a breaking change but it make more sense and I don't think it'll be a very big deal.